requirements.txt
# 旅行规划Agent依赖

# 基础依赖
requests>=2.31.0
python-dotenv>=1.0.0

# AI和大模型
openai>=1.12.0  # 用于调用千问API（兼容OpenAI格式）

# 向量数据库和知识库
chromadb>=0.4.22
sentence-transformers>=2.2.2

# 数据处理
pandas>=2.2.0
numpy>=1.26.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
